Updating notebooks html.
Updating the paths to the figures (moved to figures directory). Adding
a "next" link at the bottom of each notebook. Initially had a
"previous" link too, but that was a bad idea. The notebook server
spawns a new kernel for these notebooks so I had multiple instances
of the same notebook potentially overwriting each other (a mess).